Refactor case where getplayerspeed is 0
The goal here is to remove duplicated function calls.  This works because

v_mults(p.xthrust, 1)

is the same as p.xthrust. Also, v_mults(p.xthrust, 0) will return a zero vector. So in some sense, the code is "doing extra work" because it's calculating the player's acceleration, and velocity, and position, even though we know nothing is going to change. However, from a control flow standpoint, it's a lot clearer (at least to me).