More refactoring when getplayerspeed != 0
We can do something similar to what we did for the case where getplayerspeed is 0. That is to say, set accel to 0, 1, or -1 depending on the branch, and then calculate the acceleration, velocity, and position after the if/elseif/elseif is closed. Same for the max speed calculation. Again, we are doing "extra" work in the case where no buttons has pressed, but it doesn't really matter - accel will be zero and so it will act just like it did before.