remove separate code paths for getplayerspeed equal or not equal 0
So that last commit was kind of a setup. The point is, the code in the two mutually-exclusive code blocks are now identical. Which means we don't need to have two different code paths. Regardless of whether the player is moving or standing still, we can treat input and movement the same way.